requirement.txt
asgiref==3.9.1
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==4.2.23
django-cabinet==0.17.0
django-filer==3.3.1
django-imagefield==0.22.0
django-js-asset==3.1.2
django-mdeditor==0.1.20
django-mptt==0.17.0
django-polymorphic==4.1.0
django-tree-queries==0.20.0
easy-thumbnails==2.10
lxml==6.0.0
mysqlclient==2.2.7
pillow==11.3.0
reportlab==4.4.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
webencodings==0.5.1
